+++ lib/src/text_analyzer/_index.dart
@@ -3,7 +3,4 @@
 // to use GM Consult's proprietary software (the "Software").
 
 export 'text_analyzer.dart';
-export 'sentence.dart';
-export 'text_source.dart';
 export 'token.dart';
-export 'term_pair.dart';